feat: modularise the different functionalities
