+++ chainthru-server/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "chainthru-server"
+authors = [ "Lachezar Kolev <lachezarkolevgg@gmail.com>" ]
+description = "API server for chainthru"
+
+version.workspace = true
+edition.workspace = true
+license.workspace = true
+repository.workspace = true
+rust-version.workspace = true
+
+[dependencies]
+serde-aux.workspace = true
+serde_derive.workspace = true
+
+serde = { workspace = true, features = ["derive"]}
+sqlx = { workspace = true, features = ["postgres", "runtime-tokio"] }
+tokio = { workspace = true, features = ["full"] }
