+++ chainthru/src/lib.rs
@@ -1,0 +1,7 @@
+/// The Chainthru Ethereum Indexer.
+/// ...
+pub use chainthru_index;
+
+/// The Chainthru API server.
+/// ...
+pub use chainthru_server;
